Added support for VM sessions
Changed "sessions" to "VM sessions".
VM sessions keep track of whether the user is connected to their
VM, and when they last disconnected.
Proxy now uses the SVMP REST API to create a VM session when the
user first connects, and to update the session when the user
disconnects, so the proxy can terminate expired VMs.
Proxy runs a timeout function to kick the user when their regular
session has reached its maximum length.